# wazevo(frontend): allocation free initializeCurrentBlockKnownBounds