--- a/internal/engine/wazevo/frontend/sort_id.go
+++ b/internal/engine/wazevo/frontend/sort_id.go
@@ -1,0 +1,13 @@
+package frontend
+
+import (
+	"slices"
+
+	"github.com/tetratelabs/wazero/internal/engine/wazevo/ssa"
+)
+
+func sortSSAValueIDs(IDs []ssa.ValueID) {
+	slices.SortFunc(IDs, func(i, j ssa.ValueID) int {
+		return int(i) - int(j)
+	})
+}