--- a/internal/engine/wazevo/frontend/sort_id.go
+++ b/internal/engine/wazevo/frontend/sort_id.go
@@ -1,3 +1,5 @@
+//go:build go1.21
+
 package frontend
 
 import (