--- a/internal/engine/wazevo/frontend/sort_id_old.go
+++ b/internal/engine/wazevo/frontend/sort_id_old.go
@@ -1,0 +1,15 @@
+//go:build !go1.21
+
+package frontend
+
+import (
+	"sort"
+
+	"github.com/tetratelabs/wazero/internal/engine/wazevo/ssa"
+)
+
+func sortSSAValueIDs(IDs []ssa.ValueID) {
+	sort.SliceStable(IDs, func(i, j int) bool {
+		return int(IDs[i]) < int(IDs[j])
+	})
+}