``` 
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/stdlibs
                             │  old.txt   │             new.txt              │
                             │   sec/op   │   sec/op    vs base              │
Zig/Compile/test-opt.wasm-10   4.749 ± 2%   4.732 ± 2%       ~ (p=1.000 n=7)
Zig/Compile/test.wasm-10       5.792 ± 2%   5.798 ± 2%       ~ (p=0.383 n=7)
geomean                        5.245        5.238       -0.13%

                             │   old.txt    │              new.txt               │
                             │     B/op     │     B/op      vs base              │
Zig/Compile/test-opt.wasm-10   454.9Mi ± 0%   448.1Mi ± 0%  -1.50% (p=0.001 n=7)
Zig/Compile/test.wasm-10       703.8Mi ± 0%   694.7Mi ± 0%  -1.29% (p=0.001 n=7)
geomean                        565.8Mi        557.9Mi       -1.40%

                             │   old.txt   │              new.txt               │
                             │  allocs/op  │  allocs/op   vs base               │
Zig/Compile/test-opt.wasm-10   436.5k ± 0%   389.0k ± 0%  -10.87% (p=0.001 n=7)
Zig/Compile/test.wasm-10       555.4k ± 0%   522.0k ± 0%   -6.01% (p=0.001 n=7)
geomean                        492.4k        450.6k        -8.47%

```